requirements.txt
asgiref
certifi
charset-normalizer
cloudinary
crispy-bootstrap5
distlib
Django
django-cloudinary-storage
django-cors-headers
django-crispy-forms
django-environ
django-filter
djangorestframework
filelock
gunicorn
idna
Markdown
packaging
pillow
pipenv
platformdirs
python-dotenv
requests
setuptools
six
sqlparse
tzdata
urllib3
virtualenv